+++ libnavui-dropin/src/main/java/com/mapbox/navigation/dropin/NavigationViewStyles.kt
@@ -1,6 +1,7 @@
 package com.mapbox.navigation.dropin
 
 import android.content.Context
+import com.mapbox.maps.plugin.LocationPuck
 import com.mapbox.maps.plugin.annotation.generated.PointAnnotationOptions
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.ui.maneuver.model.ManeuverViewOptions
@@ -11,44 +12,46 @@ import kotlinx.coroutines.flow.asStateFlow
 @ExperimentalPreviewMapboxNavigationAPI
 internal class NavigationViewStyles(context: Context) {
 
-    private var _infoPanelPeekHeight: MutableStateFlow<Int> =
+    private val _infoPanelPeekHeight: MutableStateFlow<Int> =
         MutableStateFlow(ViewStyleCustomization.defaultInfoPanelPeekHeight(context))
-    private var _infoPanelMarginStart: MutableStateFlow<Int> =
+    private val _infoPanelMarginStart: MutableStateFlow<Int> =
         MutableStateFlow(ViewStyleCustomization.defaultInfoPanelMarginStart())
-    private var _infoPanelMarginEnd: MutableStateFlow<Int> =
+    private val _infoPanelMarginEnd: MutableStateFlow<Int> =
         MutableStateFlow(ViewStyleCustomization.defaultInfoPanelMarginEnd())
-    private var _infoPanelBackground: MutableStateFlow<Int> =
+    private val _infoPanelBackground: MutableStateFlow<Int> =
         MutableStateFlow(ViewStyleCustomization.defaultInfoPanelBackground())
-    private var _poiNameTextAppearance: MutableStateFlow<Int> =
+    private val _poiNameTextAppearance: MutableStateFlow<Int> =
         MutableStateFlow(ViewStyleCustomization.defaultPoiNameTextAppearance())
-    private var _tripProgressStyle: MutableStateFlow<Int> =
+    private val _tripProgressStyle: MutableStateFlow<Int> =
         MutableStateFlow(ViewStyleCustomization.defaultTripProgressStyle())
-    private var _audioGuidanceButtonParams: MutableStateFlow<MapboxExtendableButtonParams> =
+    private val _audioGuidanceButtonParams: MutableStateFlow<MapboxExtendableButtonParams> =
         MutableStateFlow(ViewStyleCustomization.defaultAudioGuidanceButtonParams(context))
-    private var _recenterButtonParams: MutableStateFlow<MapboxExtendableButtonParams> =
+    private val _recenterButtonParams: MutableStateFlow<MapboxExtendableButtonParams> =
         MutableStateFlow(ViewStyleCustomization.defaultRecenterButtonParams(context))
-    private var _cameraModeButtonParams: MutableStateFlow<MapboxExtendableButtonParams> =
+    private val _cameraModeButtonParams: MutableStateFlow<MapboxExtendableButtonParams> =
         MutableStateFlow(ViewStyleCustomization.defaultCameraModeButtonParams(context))
-    private var _routePreviewButtonParams: MutableStateFlow<MapboxExtendableButtonParams> =
+    private val _routePreviewButtonParams: MutableStateFlow<MapboxExtendableButtonParams> =
         MutableStateFlow(ViewStyleCustomization.defaultRoutePreviewButtonParams(context))
-    private var _endNavigationButtonParams: MutableStateFlow<MapboxExtendableButtonParams> =
+    private val _endNavigationButtonParams: MutableStateFlow<MapboxExtendableButtonParams> =
         MutableStateFlow(ViewStyleCustomization.defaultEndNavigationButtonParams(context))
-    private var _startNavigationButtonParams: MutableStateFlow<MapboxExtendableButtonParams> =
+    private val _startNavigationButtonParams: MutableStateFlow<MapboxExtendableButtonParams> =
         MutableStateFlow(ViewStyleCustomization.defaultStartNavigationButtonParams(context))
-    private var _speedLimitStyle: MutableStateFlow<Int> =
+    private val _speedLimitStyle: MutableStateFlow<Int> =
         MutableStateFlow(ViewStyleCustomization.defaultSpeedLimitStyle())
-    private var _speedLimitTextAppearance: MutableStateFlow<Int> =
+    private val _speedLimitTextAppearance: MutableStateFlow<Int> =
         MutableStateFlow(ViewStyleCustomization.defaultSpeedLimitTextAppearance())
-    private var _roadNameBackground: MutableStateFlow<Int> =
+    private val _roadNameBackground: MutableStateFlow<Int> =
         MutableStateFlow(ViewStyleCustomization.defaultRoadNameBackground())
-    private var _roadNameTextAppearance: MutableStateFlow<Int> =
+    private val _roadNameTextAppearance: MutableStateFlow<Int> =
         MutableStateFlow(ViewStyleCustomization.defaultRoadNameTextAppearance())
-    private var _maneuverViewOptions: MutableStateFlow<ManeuverViewOptions> =
+    private val _maneuverViewOptions: MutableStateFlow<ManeuverViewOptions> =
         MutableStateFlow(ViewStyleCustomization.defaultManeuverViewOptions())
-    private var _destinationMarkerAnnotationOptions: MutableStateFlow<PointAnnotationOptions> =
-        MutableStateFlow(ViewStyleCustomization.defaultMarkerAnnotationOptions(context))
-    private var _arrivalTextAppearance: MutableStateFlow<Int> =
+    private val _destinationMarkerAnnotationOptions: MutableStateFlow<PointAnnotationOptions> =
+        MutableStateFlow(ViewStyleCustomization.defaultDestinationMarkerAnnotationOptions(context))
+    private val _arrivalTextAppearance: MutableStateFlow<Int> =
         MutableStateFlow(ViewStyleCustomization.defaultArrivalTextAppearance())
+    private val _locationPuck: MutableStateFlow<LocationPuck> =
+        MutableStateFlow(ViewStyleCustomization.defaultLocationPuck(context))
 
     val infoPanelPeekHeight: StateFlow<Int> = _infoPanelPeekHeight.asStateFlow()
     val infoPanelMarginStart: StateFlow<Int> = _infoPanelMarginStart.asStateFlow()
@@ -76,32 +79,30 @@ internal class NavigationViewStyles(context: Context) {
     val roadNameTextAppearance: StateFlow<Int> = _roadNameTextAppearance.asStateFlow()
     val maneuverViewOptions: StateFlow<ManeuverViewOptions> = _maneuverViewOptions.asStateFlow()
     val arrivalTextAppearance: StateFlow<Int> = _arrivalTextAppearance.asStateFlow()
+    val locationPuck: StateFlow<LocationPuck> = _locationPuck.asStateFlow()
 
     fun applyCustomization(customization: ViewStyleCustomization) {
-        customization.infoPanelPeekHeight?.also { _infoPanelPeekHeight.tryEmit(it) }
-        customization.infoPanelMarginStart?.also { _infoPanelMarginStart.tryEmit(it) }
-        customization.infoPanelMarginEnd?.also { _infoPanelMarginEnd.tryEmit(it) }
-        customization.infoPanelBackground?.also { _infoPanelBackground.tryEmit(it) }
-        customization.poiNameTextAppearance?.also { _poiNameTextAppearance.tryEmit(it) }
-        customization.tripProgressStyle?.also { _tripProgressStyle.tryEmit(it) }
-        customization.recenterButtonParams?.also { _recenterButtonParams.tryEmit(it) }
-        customization.cameraModeButtonParams?.also {
-            _cameraModeButtonParams.tryEmit(it)
-        }
-        customization.routePreviewButtonParams?.also { _routePreviewButtonParams.tryEmit(it) }
-        customization.audioGuidanceButtonParams?.also {
-            _audioGuidanceButtonParams.tryEmit(it)
-        }
-        customization.endNavigationButtonParams?.also { _endNavigationButtonParams.tryEmit(it) }
-        customization.startNavigationButtonParams?.also { _startNavigationButtonParams.tryEmit(it) }
-        customization.speedLimitStyle?.also { _speedLimitStyle.tryEmit(it) }
-        customization.maneuverViewOptions?.also { _maneuverViewOptions.tryEmit(it) }
-        customization.speedLimitTextAppearance?.also { _speedLimitTextAppearance.tryEmit(it) }
+        customization.infoPanelPeekHeight?.also { _infoPanelPeekHeight.value = it }
+        customization.infoPanelMarginStart?.also { _infoPanelMarginStart.value = it }
+        customization.infoPanelMarginEnd?.also { _infoPanelMarginEnd.value = it }
+        customization.infoPanelBackground?.also { _infoPanelBackground.value = it }
+        customization.poiNameTextAppearance?.also { _poiNameTextAppearance.value = it }
+        customization.tripProgressStyle?.also { _tripProgressStyle.value = it }
+        customization.recenterButtonParams?.also { _recenterButtonParams.value = it }
+        customization.cameraModeButtonParams?.also { _cameraModeButtonParams.value = it }
+        customization.routePreviewButtonParams?.also { _routePreviewButtonParams.value = it }
+        customization.audioGuidanceButtonParams?.also { _audioGuidanceButtonParams.value = it }
+        customization.endNavigationButtonParams?.also { _endNavigationButtonParams.value = it }
+        customization.startNavigationButtonParams?.also { _startNavigationButtonParams.value = it }
+        customization.speedLimitStyle?.also { _speedLimitStyle.value = it }
+        customization.maneuverViewOptions?.also { _maneuverViewOptions.value = it }
+        customization.speedLimitTextAppearance?.also { _speedLimitTextAppearance.value = it }
         customization.destinationMarkerAnnotationOptions?.also {
-            _destinationMarkerAnnotationOptions.tryEmit(it)
+            _destinationMarkerAnnotationOptions.value = it
         }
-        customization.roadNameBackground?.also { _roadNameBackground.tryEmit(it) }
-        customization.roadNameTextAppearance?.also { _roadNameTextAppearance.tryEmit(it) }
-        customization.arrivalTextAppearance?.also { _arrivalTextAppearance.tryEmit(it) }
+        customization.roadNameBackground?.also { _roadNameBackground.value = it }
+        customization.roadNameTextAppearance?.also { _roadNameTextAppearance.value = it }
+        customization.arrivalTextAppearance?.also { _arrivalTextAppearance.value = it }
+        customization.locationPuck?.also { _locationPuck.value = it }
     }
 }