+++ libnavui-dropin/src/main/java/com/mapbox/navigation/dropin/binder/map/MapBinder.kt
@@ -1,9 +1,8 @@
 package com.mapbox.navigation.dropin.binder.map
 
 import android.view.ViewGroup
-import androidx.core.content.ContextCompat
 import com.mapbox.maps.MapView
-import com.mapbox.maps.plugin.LocationPuck2D
+import com.mapbox.maps.plugin.LocationPuck
 import com.mapbox.maps.plugin.compass.compass
 import com.mapbox.maps.plugin.locationcomponent.location
 import com.mapbox.maps.plugin.scalebar.scalebar
@@ -22,7 +21,6 @@ import com.mapbox.navigation.dropin.component.marker.MapMarkersComponent
 import com.mapbox.navigation.dropin.component.marker.RoutePreviewLongPressMapComponent
 import com.mapbox.navigation.dropin.databinding.MapboxNavigationViewLayoutBinding
 import com.mapbox.navigation.dropin.internal.extensions.reloadOnChange
-import com.mapbox.navigation.ui.app.R
 import com.mapbox.navigation.ui.app.internal.Store
 import com.mapbox.navigation.ui.app.internal.navigation.NavigationState
 import com.mapbox.navigation.ui.app.internal.routefetch.RoutePreviewAction
@@ -61,7 +59,9 @@ internal class MapBinder(
         return navigationListOf(
             CameraLayoutObserver(store, mapView, binding),
             LocationComponent(context.locationProvider),
-            locationPuckComponent(context.locationProvider),
+            reloadOnChange(context.styles.locationPuck) { locationPuck ->
+                locationPuckComponent(context.locationProvider, locationPuck)
+            },
             LogoAttributionComponent(mapView, context.systemBarsInsets),
             reloadOnChange(
                 context.mapStyleLoader.loadedMapStyle,
@@ -92,14 +92,9 @@ internal class MapBinder(
     }
 
     private fun locationPuckComponent(
-        locationProvider: NavigationLocationProvider
+        locationProvider: NavigationLocationProvider,
+        locationPuck: LocationPuck,
     ): LocationPuckComponent {
-        val locationPuck = LocationPuck2D(
-            bearingImage = ContextCompat.getDrawable(
-                mapView.context,
-                R.drawable.mapbox_navigation_puck_icon
-            )
-        )
         return LocationPuckComponent(
             mapView.getMapboxMap(),
             mapView.location,